# restaurant-system/admin/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRC Multicuisine Restaurant Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthPayload"
        "401":
          description: Invalid credentials
...
  /offers:
    get:
      operationId: getOffers
      tags: [offers]
      summary: Get all offers
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Offers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Analytics
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get all notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthPayload:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/Admin"
    Admin:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]

    # Dashboard
...
    Offer:
      type: object
      required: [id, title, originalPrice, discountedPrice, discountPercent, isActive, startDate, endDate]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        originalPrice:
          type: number
        discountedPrice:
          type: number
        discountPercent:
          type: number
        badge:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        isFeatured:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [title, originalPrice, discountedPrice, discountPercent, startDate, endDate]
      properties:
        title:
          type: string
        description:
          type: string
        bannerUrl:
          type: string
        originalPrice:
          type: number
        discountedPrice:
          type: number
        discountPercent:
          type: number
        badge:
          type: string
        category:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isFeatured:
          type: boolean
        isActive:
          type: boolean
    OfferUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        bannerUrl:
          type: string
        originalPrice:
          type: number
        discountedPrice:
          type: number
        discountPercent:
          type: number
        badge:
          type: string
        category:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isFeatured:
          type: boolean
        isActive:
          type: boolean

    # Analytics
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [new_order, cooking_delay, new_review, offer_expired, revenue_milestone, table_update]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        relatedId:
          type: ["integer", "null"]
        createdAt:
          type: string
...